Add the --bind option in "docker build"
The --bind option is an optimisation for the "docker build" command.

Its purpose is to avoid uploading the context when the client is
connecting from the same host as the server. Instead the client just
sends the path where its context is located and the server binds it to
the temporary location where the server-side context will reside
(/tmp/docker-build-xxx).

Together with a8cf120283f8bbfe04aff323d89251816b043a99 this patch
enables very quick iterations when dealing with large contextes
(it reduces the actual number of copies of the context to 0 instead of 3
when the Dockerfile does "ADD . /somewhere"). It is an efficient
workaround for #2224.

Implementation notes:
 - to avoid altering the client orig context, the bind is set up through
   aufs, using an additional layer to store temporary server-side
   modifications

 - to enable caching intermediate images but without sha256summing every
   file (which would ruin the optimisation), the fingerprint is only
   made from the metadata from stat(2); thus rebuilds are triggered when
   mtime changes (like in good old makefiles)

WARNING: in its current form, this patch does not check any credential
         (the client can mount any directory from the server,
         giving full read access to its filesystem). DO NOT USE IT if
         ordinary/external users are allowed to connect to the docker
         socket.